refactor: 검색/알림 컴포넌트 분리
